cache-auto-updater/requirements.txt
functions-framework==3.0.0
google-cloud-storage==2.13.0
google-genai==0.3.0
